编写生成字节码结构框架

+++ core/libgrammar/src/lexical/number/mod.rs
@@ -349,11 +349,11 @@ impl<T: FnMut() -> CallbackReturnStatus, CB: Grammar> LexicalParser<T, CB> {
         if value >= u8::min_value() as u64 && value <= u8::max_value() as u64 {
             return (PrimevalType::Uint8, PrimevalData::Uint8(Some(number::uint8::Uint8::new(value as u8))));
         } else if value > u8::max_value() as u64 && value <= u16::max_value() as u64 {
-            return (PrimevalType::Uint8, PrimevalData::Uint16(Some(number::uint16::Uint16::new(value as u16))));
+            return (PrimevalType::Uint16, PrimevalData::Uint16(Some(number::uint16::Uint16::new(value as u16))));
         } else if value > u16::max_value() as u64 && value <= u32::max_value() as u64 {
-            return (PrimevalType::Uint8, PrimevalData::Uint32(Some(number::uint32::Uint32::new(value as u32))));
+            return (PrimevalType::Uint32, PrimevalData::Uint32(Some(number::uint32::Uint32::new(value as u32))));
         } else {
-            return (PrimevalType::Uint8, PrimevalData::Uint64(Some(number::uint64::Uint64::new(value))));
+            return (PrimevalType::Uint64, PrimevalData::Uint64(Some(number::uint64::Uint64::new(value))));
         }
     }
